+++ core/e2e/reverse_interface.go
@@ -1,8 +1,9 @@
+//go:generate go run gopkg.d7z.net/go-mini/cmd/ffigen -pkg e2e -out reverse_ffigen_test.go reverse_interface.go
 package e2e
 
 // ffigen:reverse
 type ScriptCalculator interface {
-	Add(a, b int) int
-	Format(prefix string, val int) string
-	Divide(a, b int) (int, error)
+	Add(a, b int64) int64
+	Format(prefix string, val int64) string
+	Divide(a, b int64) (int64, error)
 }
